chore: consolidacao de todas as mudancas pendentes da sessao
Inclui:
- HikariCP migration (substituicao C3P0 por HikariCP 5.1)
- Remocao de libs legadas (c3p0, mchange-commons, mysql-connector 5.1, coroutines 1.1.1)
- Kotlin deprecation fixes (diversos arquivos)
- Gradle wrapper + settings.gradle.kts
- Database factory modernizado (L2DatabaseFactory.kt)
- Fixes em quests, skills, AI, handlers, network packets
- Login packet adjustments
- Config hexid.txt atualizado

+++ src/com/l2kt/gameserver/skills/l2skills/L2SkillSiegeFlag.kt
@@ -35,7 +35,7 @@ class L2SkillSiegeFlag(set: StatsSet) : L2Skill(set) {
         npcDat.set("name", player!!.clan.name)
         npcDat.set("usingServerSideName", true)
 
-        npcDat.set("hp", if (_isAdvanced) 100000 else 50000)
+        npcDat.set("hp", (if (_isAdvanced) 100000 else 50000) as Int)
         npcDat.set("mp", 0.toInt())
 
         npcDat.set("radius", 10.toInt())
